fix cleanup errors during provisioning
When provisioning fails and the deferred cleanup function is called, the
controller must take care to ensure the object bucket resource being
cleaned up (deleted) has the necessary fields filled in for delete calls
to be successful. Fill in basic object bucket info as soon as possible
after Privision() or Grant() returns, and in the event of an error
during one of these calls, make sure the deferred cleanup function also
sets the information it can before calling Delete().

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>