fix errors during provisioning (#198)
* fix cleanup errors during provisioning

When provisioning fails and the deferred cleanup function is called, the
controller must take care to ensure the object bucket resource being
cleaned up (deleted) has the necessary fields filled in for delete calls
to be successful. Fill in basic object bucket info as soon as possible
after Privision() or Grant() returns, and in the event of an error
during one of these calls, make sure the deferred cleanup function also
sets the information it can before calling Delete().

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>

* fix code generation

The code generation script was not up to date for the k8s v1.19 API.
Update code gen to work again, and update the generated code.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>

* allow provisioner to override OB reclaim policy

Do not blindly fill in or overwrite the reclaim policy of an object
bucket. A provisioner implementation might have reason to specify a
reclaim policy for an object bucket that is different from what the
storage class specifies.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>