Recover from provision case with existing OB
If an implementing Provisioner fails to include "objectBucketName" in
the CRDs needed for object bucket claims, Kubernetes may drop the
unknown "objectBucketName" parameter that this library uses to determine
if an OBC is bound to an OB. Allow this library to recover from this
case or other degraded cases that are similar by checking whether the
Object Bucket that would be created for the claim already exists.

If the OB exists during provisioning and has an owner reference to the OBC,
then do not continue provisioning. Instead, attempt to update the OBC to
refer to the OB via "objectBucketName", and update the phase to "Bound".

Another unlikely but possible case where this degraded case might be
encountered if the Provisioner operator crashes after updating the OB but
before the OBC can be updated. The changes here will also help recover
from this edge case.

When provisioning an OBC, the library can fail on creation of secret or
configmap in the case where the operator crashes after secret and/or
configmap are created but before provisioning is finished. Allow the
provision code to find existing secrets and/or configmaps and continue
with provisioning if they exist rather than erroring out to infinitely
fail provisioning in a loop.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>