Fix OB empty check during provisioning
Fix the existing check for an OB empty struct by using k8s api's semantic
deep-equality methods. Previously, this was a pointer equality check and
not a valid check for an empty struct. The k8s code generator does not
generate `==` or other equality methods for generated types that might
have been able to be used.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>