fix OBC delete stuck
The delete operation is handled via update calls, with addition of new
Update() API, the objects need to deleted are not added to the queue due
to check in updateSupported(). So if the deletion timestamp is set, then
this check is ignored in NewController().
FInd more details [here](https://github.com/rook/rook/pull/8514#issuecomment-898011184)

Signed-off-by: Jiffin Tony Thottan <thottanjiffin@gmail.com>