BREAKING! allow OBs to be updated by provisioner
THIS IS A BREAKING CHANGE!

Sometimes, an OB might need to be updated by the provisioner. For
example, the endpoint of the object store changed.

In order for this to work, lib-bucket-provisioner must use an idempotent
controller, and it must call Provision()/Grant() on each reconcile.
Provision() and Grant() must then also be idempotent. In order to more
easily allow these calls to be implemented idempotently,
lib-bucket-provisioner will now call a new GenerateUserID() function
that is expected to deterministically create a UserID that will be used
in subsequent calls to Provision() and Grant() for use as an idempotency
key.

Because Provision() and Grant() are idempotent, there is now no need for
the Update() method.

This is a chagne to the previous API spec which allowed provisioners to
return errors if a bucket already existed in the backend.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>